Refactor ChannelBuilder to be in a helpers file
